requirements.txt
pinecone>=3.0.0
openai>=1.0.0
anthropic>=0.25.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
tqdm>=4.66.0
flask>=3.0.0
gunicorn>=21.0.0
